generative_ai/pyproject.toml
[tool.poetry]
name = "generative_ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
retrieval-agent-fireworks = {path = "packages\\retrieval-agent-fireworks", develop = true}
langchain-pinecone = "^0.1.0"
lxml = "^5.2.1"
bs4 = "^0.0.2"
langchain-cohere = "^0.1.4"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
